==> Implementations/FY20/ACC_Peru_Lima_Optimization/JupyterNotebooks/Step3 - Data preparation and OD matrix computation.ipynb <==
### Prepare road network, snap origins and destination points to road network nodes, and compute OD
general steps:
1. add time to edges
2. take only biggest subgraph
3. snap origins and destinations to the road network
4. calculate the OD (origin-destination matrix)

import geopandas as gpd
import os, sys, time
import pandas as pd
sys.path.append(r'../../../GOSTNets/GOSTNets')
import GOSTnet as gn
import importlib
import networkx as nx
import osmnx as ox
from shapely.ops import unary_union
from shapely.wkt import loads
from shapely.geometry import LineString, MultiLineString, Point

Prepare data for the OD matrix:
    1. Create your origins (point) .csv file - be sure to name the column of the unique identifier with "O_ID" and save the coordinates into columns "Lat" / "Lon" 
    2. Create your destinations (point) .csv file - be sure to save the coordinates in columns named "Lat" / "Lon" 

pth = r'../../Peru_Lima_Optimization/test_data'
OD_path = r'../../../../lima_optimization_output'
graph_path = r'../../../../lima_optimization_output'
graph_file = r'Lima_processed.pickle'
origins_file = r'origins_VES_CentroidUrbanBlocks.csv' #change with your .csv file name
destinations_file = r'destinations_VES_HealthFacilities.csv'  #change with your .csv file name

#### 1. Read in road network

G = nx.read_gpickle(os.path.join(graph_path, graph_file))

G.number_of_edges()

G.number_of_nodes()

The first cell contains the standard speed limits for each road type from OSM. However, it is recommended to adjust these values if the AOI has specific travel conditions (see below for Lima, Peru).

# default speed limits from OSM

#sd = {          'residential': 20,  # kmph
#                'primary': 40, # kmph
 #               'primary_link':35,
  #              'motorway':35,
   #             'motorway_link': 25,
    #            'trunk': 20,
     #           'trunk_link':15,
      #          'secondary': 10, # kmph
       #         'secondary_link':5,
        #        'tertiary':5,
         #       'tertiary_link': 5,
          #      'unclassified':5
           #     }

#speed limits for Lima
sd = {          'residential': 10,  # kmph
                'primary': 25, # kmph
                'primary_link':20,
                'motorway':35,
                'motorway_link': 25,
                'trunk': 20,
                'trunk_link':15,
                'secondary': 10, # kmph
                'secondary_link':5,
                'tertiary':5,
                'tertiary_link': 5,
                'unclassified':5
                }

gn.example_edge(G)

### 1.) add time to edges
Here you can change the graph_type from "drive" to "walk" if needed. Instead of driving time this will compute walking time using the standard walk_speed = 4.5 km/h (this can be modified in the GOSTnets.py file) 

G_time = gn.convert_network_to_time(G, 
                                   distance_tag = 'length',
                                   graph_type = 'drive', 
                                   road_col = 'infra_type',
                                   speed_dict = sd, 
                                   factor = 1000)

gn.example_edge(G_time)

### 2.) take only biggest routable subgraph from the road network
This is done because disconnected networks are not routable

D = list(nx.strongly_connected_component_subgraphs(G_time))

G = D[0]

G = nx.convert_node_labels_to_integers(G)

gn.save(G, 'biggest_subg', G_path)

### 3.) snap origins and destinations to the road network
origins and destination nodes need to be snapped directly to the road network nodes in order to do routing

O = pd.read_csv(os.path.join(pth, origins_file), sep=',', encoding = "ISO-8859-1")
D = pd.read_csv(os.path.join(pth, destinations_file), sep=',', encoding = "ISO-8859-1")

O[:3]

D[:3]

#### The function pandana_snap maps each origin and destination point to the nearest node on the road graph. It will create a GeoPandasDataFrame and it will add a column titled 'NN' that has the nearest road network node ID for each snapped point. If 'add_dist_to_node_col = True' then another column titled 'NN_dist' will be added that will have the distance from each snapped point to its nearest road network node.

O['geometry'] = list(zip(O['Lon'],O['Lat']))
O['geometry'] = O['geometry'].apply(Point)
O_gdf = gpd.GeoDataFrame(O, crs = {'init':'epsg:4326'}, geometry = 'geometry')

O_gdf[:5]

# Be sure to input the correct local utm zone for the target_crs in order to have the correct distance computed (in meters)
O_gdf = gn.pandana_snap(G, O_gdf, target_crs = 'epsg:32718', add_dist_to_node_col = True)
origins = list(O_gdf.NN)
origins = list(set(origins))

O_gdf[:3]

D['geometry'] = list(zip(D['Lon'],D['Lat']))
D['geometry'] = D['geometry'].apply(Point)
D_gdf = gpd.GeoDataFrame(D, crs = {'init':'epsg:4326'}, geometry = 'geometry')
D_gdf = gn.pandana_snap(G, D_gdf, target_crs = 'epsg:32718', add_dist_to_node_col = True)
destinations = list(D_gdf.NN)
destinations = list(set(destination))

print("length of origins is %s" % len(origins))

print("length of destinations is %s" % len(destinations))

Write the origins_snapped and destinations_snapped files.

O_gdf.to_csv(os.path.join(OD_path, 'origins_snapped.csv'))
D_gdf.to_csv(os.path.join(OD_path, 'destinations_snapped.csv'))

### 4.) calculate the OD (origin-destination matrix)

run a simulation of OD generation below, with a subset of 50 origins and 50 destinations - a good way to estimate the time it will take to run the entire OD

o_test = origins[:10]
print(o_test)

d_test = destinations[:10]
print(d_test)

***The gn.calculate_OD function enables the users to choose a computation between "time" and "distance" between each origin and destination point. ***

%time OD_test = gn.calculate_OD(G, o_test, d_test, fail_value = 9999999999999, weight = 'time')

This is the matix dimension of your whole dataset

len(origins),len(destinations)

# count how large your OD matrix is
len(origins) * len(destinations)

# do real OD calculation
%time OD = gn.calculate_OD(G, origins, destinations, fail_value = 9999999999999, weight = 'time')

OD_df = pd.DataFrame(OD, columns = destinations , index = origins)

OD_df[:5]

#Write OD to file
OD_df.to_csv(os.path.join(OD_path,'saved_OD.csv'))

***** Depending on the size of the problem you are trying to solve, you might consider running the OD computation using the graph-tool python module in a virtual environment*****
