==> Implementations/FY20/ACC_Peru_Lima_Optimization/JupyterNotebooks/Step4 - Accessibility analysis - data prep and mapping of the OD.ipynb <==
### Map results back to shapefile
The functions below prepare the data for the process of mapping the results back to their original shapefiles by implementing a "table join" between the outputs .csv and the attribute table of the shapefile. This will allow the visualization of accessibility **before** optimization.

import pandas as pd
import os, sys
sys.path.append(r'../../../GOSTNets/GOSTNets')
import GOSTnet as gn
import importlib
importlib.reload(gn)
import geopandas as gpd
from shapely.wkt import loads
import numpy as np

#define the paths to the origins, destinations and .pickle file
pth = r'../../../../lima_optimization_output'
WGS = {'init':'epsg:4326'}
measure_crs = {'init':'epsg:32718'}

# read  file
OD_name = r'saved_OD.csv'
OD = pd.read_csv(os.path.join(pth, OD_name), sep=',')
OD = OD.rename(columns = {'Unnamed: 0':'NN'})

#preview
OD[:3]

OD_ = OD.set_index('NN')
#replacing infinity and negative infinity with not a number
OD = OD_.replace([np.inf, -np.inf], np.nan)
#adding a column that has the minimum time to reach each origin's nearest facility
OD['Minimum'] = OD.loc[:,:].min(axis=1)

OD[:3]

### PREPARE ORIGIN FILE 
set index to nearest node to match on the travel time from that node to nearest destination (OD['Minimum'])

orig = pd.read_csv(os.path.join(pth, 'origins_snapped.csv'))
#orig = orig.rename({'ID':'OBJECTID'}, axis = 1) # rename ID column to 'OBJECTID'
orig = orig.set_index('NN') # used for matching on the OD matrix
orig[:3]

#join works on the index as default 
join = orig.join(OD)
join

# add on walktime (kph) for time computation from each origin node's original location to its nearest node on the graph
walkspeed = 4
# generate walktime in seconds to 'NN'
join['walktime'] = join['NN_dist'] / 1000 / walkspeed * 3600
# total_time in minutes
join['total_time_min'] = (join['walktime'] + join['Minimum']) / 60
###   OR   ###
# compute total distance: NN_dist + dist
# this is assuming that the 'Minimum' values are distances instead of time
#join['total_distance'] = (join['NN_dist'] + join['Minimum'])

#read in the original origins (neighborhood polygons) in shapefile format
fp = r'../../Peru_Lima_Optimization/test_data'
shpfil = gpd.read_file(os.path.join(fp, 'VillaElSalvador_urban.shp'))

# Ensure both shapefile AND origin file have a non-duplicate index to enable matching of results on to the shapefile
if shpfil.index.name != 'OBJECTID':
    shpfil = shpfil.set_index('OBJECTID')
#join.index.names = ['OBJECTID']

shpfil[:10]

# testing
#join
#join.loc[join['O_ID'] < 10]

# set index to O_ID so that you can add the 'total_time' column to shpfil GPD, which also has an index of the Object ID
join = join.set_index('O_ID')

# Prep output shapefile, match on visualization column, save file
shpfil['total_time_min'] = join['total_time_min']

shpfil[:5]

shpfil = shpfil.to_crs({'init':'epsg:4326'})
shpfil.to_file(os.path.join(pth, 'VES_origins_w_time_pre_optimization.shp'), driver = 'ESRI Shapefile')

# write the road file as a shapefile
from shapely.wkt import loads
edges = pd.read_csv(os.path.join(pth, 'biggest_subg_edges.csv'))
edges['geometry'] = edges['geometry'].apply(loads)
edges_gdf = gpd.GeoDataFrame(edges, geometry = 'geometry', crs = {'init':'epsg:4326'})
edges_gdf.to_file(os.path.join(pth, 'VES_roads.shp'), driver = 'ESRI Shapefile')
