==> Implementations/FY20/ACC_Peru_Lima_Optimization/JupyterNotebooks/Step5 - Data preparation before optimization (optional).ipynb <==
### Slice OD as needed; compute weighted OD
In this notebook we have the opportunity to adjust the inputs for the optimization analysis:
- a choice can be made on the number and type of destinations that will be considered in the process
- a choice can be made between using a weighted or non-weighted OD matrix (in this example we compute a weighted OD by population )

import geopandas as gpd
import os, sys, time
import pandas as pd
sys.path.append(r'../../../GOSTNets/GOSTNets')
import GOSTnet as gn
import importlib
import networkx as nx
import osmnx as ox
from shapely.ops import unary_union
from shapely.wkt import loads
from shapely.geometry import LineString, MultiLineString, Point

#define the paths to the origins, destinations files and the OD
pth = r'../../../../lima_optimization_output'

 ### Step A) Filter destinations 
 Run this step if you need to filter the OD by a subset of the destination file

#read destinations snapped csv (from Step3)
dest_snapped_df = pd.read_csv(os.path.join(pth, 'destinations_snapped.csv'))
dest_snapped_df[:3]

#filter destinations file by hospital category 
hour_12_dest_snapped_df = dest_snapped_df.loc[(dest_snapped_df['categoria'] == 11) | (dest_snapped_df['categoria'] == 12)]
hour_12_dest_list = hour_12_dest_snapped_df.NN.tolist()
len(hour_12_dest_list)

hour_12_dest_list

#### This list will eventually be used at a later step to only keep the columns (destinations) in the OD matrix than match this list 

#write new destiantion file to .csv
hour_12_dest_snapped_df.to_csv(os.path.join(pth,'hour_12_dest_snapped_df.csv'))

### Step B)
Run this step if you want to compute a weighted OD

#read origins snapped csv
origins_snapped_df =  pd.read_csv(os.path.join(pth, 'origins_snapped.csv'))
origins_snapped_df[:3]

#create a series of the unique Nearest Node (NN) and its corresponding population value
origins_w_demands_series = pd.Series(origins_snapped_df.Population.values,index=origins_snapped_df.NN)
len(origins_w_demands_series)

#some origins end up snapping to the same nearest node, therefore the code below groups and sums origin populations
origins_no_dupl = origins_w_demands_series.groupby('NN').sum()
len(origins_no_dupl)

#### Generate Weighted OD Matrix

#load OD and multiply OD by population weight -> weighted OD
OD_df = pd.read_csv(os.path.join(pth, 'saved_OD.csv'), index_col=0, sep=',')
OD_df[:3]

OD_weighted_df = OD_df.mul(origins_no_dupl, axis=0)
OD_weighted_df[:3]

## Step C)
This step is done if you have a weighted OD and then you wanted to filter it with the list you generated in Step A

hour_12_dest_list

# it might be necessary to transform the OD entries into type "int"
OD_weighted_df.dtypes
OD_weighted_df_int = OD_weighted_df.astype(int)
OD_weighted_df_int[:3]

# transform the destination nodes into "string" type to be used as column headers
hour_12_dest_list_s = [str(x) for x in hour_12_dest_list]
hour_12_dest_list_s

# filter the OD by the destination -category 12
OD_df_12_hr = OD_weighted_df_int[hour_12_dest_list_s]
OD_df_12_hr[0:10]

# save new OD as a CSV, this will be an input into optimization analysis
OD_df_12_hr.to_csv(os.path.join(pth,'OD_weighted_df_12_hr.csv'))
