==> Implementations/FY20/ACC_Peru_Lima_Optimization/JupyterNotebooks/Step6 - Optimization_Max_Covering_Python.ipynb <==
This is a type of location optimization analysis, specifically finding the optimal location of facilites on a network. These are two types of Set-Coverage analysis implemented in **Python**:

### The Maximal Covering Location Problem (MCLP)
#### Objective: The Maximal Covering Location Problem determines the location of P facilities in order to maximize the demand covered within a pre-specified maximum distance coverage.

more information on GOSTNets Optimization can be found in the wiki: https://github.com/worldbank/GOST_PublicGoods/wiki/GOSTnets-Optimization

import geopandas as gpd
import pandas as pd
import os, sys, time
sys.path.append(r'../../../GOSTNets/GOSTNets')
import GOSTnet_Optimization as gn
import importlib
import networkx as nx
import osmnx as ox

pth = r'../../../../lima_optimization_output'

### load OD matrix

OD_df = pd.read_csv(os.path.join(pth, 'saved_OD.csv'),index_col=0)

### read origins snapped csv

origins_snapped_df =  pd.read_csv(os.path.join(pth, 'origins_snapped.csv'))
origins_snapped_df[:3]

#create a series of the unique Nearest Node (NN) and its corresponding population value
origins_w_demands_series = pd.Series(origins_snapped_df.Population.values,index=origins_snapped_df.NN)
len(origins_w_demands_series)

#some origins end up snapping to the same nearest node, therefore the code below groups and sums origin populations
origins_w_demands_series_no_dupl = origins_w_demands_series.groupby('NN').sum()
len(origins_w_demands_series_no_dupl)

### we produced a series that has each origin and its respective population

origins_w_demands_series_no_dupl

#for Julia functions
origins_w_demands_series_no_dupl.to_csv(os.path.join(pth, 'origins_w_demands_series_no_dupl.csv'), header=["pop"])

### Inputs include the OD matrix, locating 3 facilities, a pre-specified maximum distance coverage of 2000 seconds, and a dictionary of origins with their population

import importlib
importlib.reload(gn)

max_coverage_result = gn.optimize_max_coverage(OD_df, p_facilities = 3, max_coverage = 200, origins_pop_series = origins_w_demands_series_no_dupl, existing_facilities = None)

max_coverage_result
