==> Implementations/FY20/ACC_Peru_Lima_Optimization/JupyterNotebooks/Step6 - Optimization_P-Median_Python.ipynb <==
This is a type of location optimization analysis, specifically finding the optimal location of facilites on a network. This analysis is the P-Median Problem implemented in **Python**:

### P-Median Problem
The P-median problem finds the location of (a pre-specified number of) P facilities to minimize the average travel distance (or time) among all demand points and facilities. The P-median problem can take into account the level of demand at each point (e.g. number of people, or the number of visits).

more information on GOSTNets Optimization can be found in the wiki: https://github.com/worldbank/GOST_PublicGoods/wiki/GOSTnets-Optimization

import geopandas as gpd
import pandas as pd
import os, sys
# add to your system path the location of the GOSTnet.py library
sys.path.append(r'../../../GOSTNets/GOSTNets')
import GOSTnet_Optimization as gn
import importlib

## The P-Median Problem requires an OD matrix as an input
### read in OD matrix saved as csv file as dataframe

pth = r'../../../../lima_optimization_output'
OD_df = pd.read_csv(os.path.join(pth, 'saved_OD.csv'),index_col=0)

OD_df[:3]

OD_df.keys()

facilities = OD_df.columns.values.tolist()

# facilities list
facilities

# pulp is a python optimization library
import pulp
#import importlib
#importlib.reload(gn)

## Compute the P-Median Problem
The third argument to the optimize_facility_locations function is the number facilities to optimally locate

results = gn.optimize_facility_locations(OD_df, facilities, 4, existing_facilities = None)

### These are the node IDs of the selected optimal facilities

results

# save results in a .txt file
file = open(r'../../../../lima_optimization_output/results.txt', 'w')
for r in results:
    file.write('{}\n'.format(r))
file.close()

### Optional, if you would like to access the objective function value

#generate pulp.LpProblem object without solving
return_problem = gn.optimize_facility_locations(OD_df, facilities, 4, existing_facilities = None, execute=False)

#solve problem
return_problem.solve()

#print out objective value
return_problem.objective.value()
