==> Implementations/FY20/ACC_Peru_Lima_Optimization/JupyterNotebooks/Step6 - Optimization_SetCoverage_Julia.ipynb <==
This is a type of location optimization analysis, specifically finding the optimal location of facilites on a network. These are two types of Set-Coverage analysis implemented in **Julia**:

### Set-Coverage Problem
#### Objective: Determine the minimum number of facilities and their locations in order to cover all demands within a pre-specified maximum distance (or time) coverage

### Partial Set-Coverage Problem
#### Objective: Determine the minimum number of facilities and their locations in order to cover a given fraction of the population within a pre-specified maximum distance (or time) coverage

more information on GOSTNets Optimization can be found in the wiki: https://github.com/worldbank/GOST_PublicGoods/wiki/GOSTnets-Optimization

#### This is a Julia Notebook. If you are new to Julia, these are the [steps](https://datatofish.com/add-julia-to-jupyter/) to add Julia to a Jupyter Notebook

using Pkg
Pkg.add("JuMP")
Pkg.add("Cbc")
Pkg.add("MathOptInterface")
Pkg.add("MathProgBase")
Pkg.add("CSV")
Pkg.add("DelimitedFiles")
Pkg.add("DataFrames")
println("Done installing packages")

#using JuMP, Cbc, GLPK, CPLEX, Test, Random, MathOptInterface, MathOptFormat, CSV, DataFrames, DelimitedFiles, MathProgBase
using JuMP, Cbc, MathOptInterface, CSV, DataFrames, DelimitedFiles, MathProgBase

# MathOptInterface is an abstraction layer for mathematical optimization solvers
const MOI = MathOptInterface

## This is the Julia Set Coverage function

function SetCoverageProblem(CSVfile, maxCoverage::Int)

    println("maxCoverage value")
    println(maxCoverage)

    # materialize a csv file as a DataFrame
    df = CSV.File(CSVfile) |> DataFrame!

    #extract column_headers
    column_headers = []
    #skip Column1
    for i=2:length(names(df))
      push!(column_headers,String(names(df)[i]))
    end
    
    OD_dict = Dict()
    for i in 1:size(df, 1)
        OD_dict[df[i,1]] = df[i,2:end]
    end

    #println("print OD_dict")
    #println(OD_dict)

    #origins as array
    origins = df[:,1]

    println("origins")
    println(origins)

    facilities = []
    for i in df[1,2:end]
      push!(facilities,trunc(Int, i))
    end

    println("facilities")
    println(facilities)

    #m = Model(with_optimizer(CPLEX.Optimizer))
    #output says threads were changed, but I do not see a difference on the resource monitor
    #m = Model(with_optimizer(Cbc.Optimizer, threads = 14))
    #change the limit to 
    m = Model(with_optimizer(Cbc.Optimizer, threads = 2, seconds = 68400))

    # Facility locations
    #@variable(m, 0 <= s[1:numLocation] <= 1)
    #@variable(m, 0 <= x[1:length(facilities)] <= 1)
    #binary variable
    @variable(m, x[1:length(facilities)], binary=true)

    #println("print Facility location var")
    #println(x)

    # Aux. variable: x_a,i = 1 if the closest facility to a is at i
    #@variable(m, 0 <= x[1:numLocation,1:numCustomer] <= 1)
    #@variable(m, 0 <= y[origins,1:length(facilities)] <= 1)
    #binary variable
    #@variable(m, y[origins,1:length(facilities)], binary=true)

    #println("print origin facility var")
    #println(y)

    # Objective: min distance
    #@objective(m, Min, sum(abs(customerLocations[a]-i)*x[i,a] for a = 1:numCustomer, i = 1:numLocation) )

    
    #println("testing1")
    #for j in facilities
    #    println(j)
    #end
    
    #@objective(m, Min, sum(OD_dict[i][j]*y[i,j] for i in origins, j = 1:length(facilities)) )
    @objective(m, Min, sum(x[j] for j in 1:length(facilities)))
    
    # Constraints


    # Subject to must allocate all facilities
    #@constraint(m, sum(x[i] for i=1:length(facilities)) == numFacility )


    for i in origins
        eligibleFacilities = []
        for j in 1:length(facilities)
            if OD_dict[i][j] <= maxCoverage
                push!(eligibleFacilities,j)
            end
        end
        @constraint(m, sum(x[j] for j in eligibleFacilities) >= 1 )
    end


    JuMP.optimize!(m)

    println("Objective value is: ", JuMP.objective_value(m))

    #println("Objective bound is: ", JuMP.objective_bound(m))


    println("print array values")
    println(value.(x))
    println("print array length")
    println(length(value.(x)))

    result_array = value.(x)

    selected_facilities = []

    for i=1:length(result_array)
       if result_array[i] == 1
           push!(selected_facilities,column_headers[i])
       end
    end

    println("print selected_facilities")
    println(selected_facilities)



    if termination_status(m) == MOI.OPTIMAL
        optimal_solution = value.(x)
        optimal_objective = objective_value(m)
    elseif termination_status(m) == MOI.TIME_LIMIT && has_values(model)
        suboptimal_solution = value.(x)
        suboptimal_objective = objective_value(m)
    else
        error("The model was not solved correctly.")
    end

    return selected_facilities

end

### The SetCoverage function takes in the OD matrix as a csv file for the first argument, and for the second argument it takes in the max coverage value.

selected_facilities = SetCoverageProblem("../../../../lima_optimization_output/saved_OD.csv", 1200)

selected_facilities

## This is the Partial Julia Set Coverage function

function PartialSetCoverageProblem(CSVfile, maxCoverage::Int, popCoverage, origins_pop_dict)

    println("maxCoverage value")
    println(maxCoverage)

    # materialize a csv file as a DataFrame
    df = CSV.File(CSVfile) |> DataFrame!

    #extract column_headers
    column_headers = []
    #skip Column1
    for i=2:length(names(df))
      push!(column_headers,String(names(df)[i]))
    end
    
    OD_dict = Dict()
    for i in 1:size(df, 1)
        OD_dict[df[i,1]] = df[i,2:end]
    end

    #println("print OD_dict")
    #println(OD_dict)

    #origins as array
    origins = df[:,1]

    #println("origins")
    #println(origins)

    facilities = []
    for i in df[1,2:end]
      push!(facilities,trunc(Int, i))
    end

    println("facilities")
    println(facilities)

    #m = Model(with_optimizer(CPLEX.Optimizer))
    #output says threads were changed, but I do not see a difference on the resource monitor
    #m = Model(with_optimizer(Cbc.Optimizer, threads = 14))
    #change the limit to 
    m = Model(with_optimizer(Cbc.Optimizer, threads = 2, seconds = 68400))

    # Facility locations
    #@variable(m, 0 <= s[1:numLocation] <= 1)
    #@variable(m, 0 <= x[1:length(facilities)] <= 1)
    #binary variable
    @variable(m, x[1:length(facilities)], binary=true)

    #println("print Facility location var")
    #println(x)

    # Aux. variable: x_a,i = 1 if the closest facility to a is at i
    #@variable(m, 0 <= x[1:numLocation,1:numCustomer] <= 1)
    #@variable(m, 0 <= y[origins,1:length(facilities)] <= 1)
    #binary variable
    @variable(m, z[origins], binary=true)

    #println("print origin facility var")
    #println(y)

    # Objective: min distance
    #@objective(m, Min, sum(abs(customerLocations[a]-i)*x[i,a] for a = 1:numCustomer, i = 1:numLocation) )

    
    #println("testing1")
    #for j in facilities
    #    println(j)
    #end
    
    #@objective(m, Min, sum(OD_dict[i][j]*y[i,j] for i in origins, j = 1:length(facilities)) )
    @objective(m, Min, sum(x[j] for j in 1:length(facilities)))
    
    # Constraints


    # Subject to must allocate all facilities
    #@constraint(m, sum(x[i] for i=1:length(facilities)) == numFacility )


    for i in origins
        eligibleFacilities = []
        for j in 1:length(facilities)
            if OD_dict[i][j] <= maxCoverage
                push!(eligibleFacilities,j)
            end
        end
        @constraint(m, sum(x[j] for j in eligibleFacilities) - z[i] >= 0 )
    end


    # if origins_pop_dict is not NONE
    
    origins_pop_dict_sum = 0
    for (key,value) in origins_pop_dict
        origins_pop_dict_sum = origins_pop_dict_sum + value
        #println(value)
    end
    println("print origins_pop_dict_sum")
    println(origins_pop_dict_sum)
    
    min_coverage = origins_pop_dict_sum * popCoverage

    println("print min_coverage")
    println(min_coverage)
    
    @constraint(m, sum(z[i] * origins_pop_dict[i] for i in origins) >= min_coverage)
    
    JuMP.optimize!(m)

    println("Objective value is: ", JuMP.objective_value(m))

    #println("Objective bound is: ", JuMP.objective_bound(m))


    println("print array values")
    println(value.(x))
    println("print array length")
    println(length(value.(x)))

    result_array = value.(x)

    selected_facilities = []

    for i=1:length(result_array)
       if result_array[i] == 1
           push!(selected_facilities,column_headers[i])
       end
    end

    println("print selected_facilities")
    println(selected_facilities)



    if termination_status(m) == MOI.OPTIMAL
        optimal_solution = value.(x)
        optimal_objective = objective_value(m)
    elseif termination_status(m) == MOI.TIME_LIMIT && has_values(model)
        suboptimal_solution = value.(x)
        suboptimal_objective = objective_value(m)
    else
        error("The model was not solved correctly.")
    end

    return selected_facilities

end

### Inputs include the OD matrix, a pre-specified maximum distance coverage, a pop coverage value, and a dictionary of origins with their population  

### import an origins_pop_series from csv

# materialize a csv file as a DataFrame
origins_pop_series = CSV.File("../../../../lima_optimization_output/origins_w_demands_series_no_dupl.csv") |> DataFrame!

origins_pop_dict = Dict()
for i in 1:size(origins_pop_series,1)
    origins_pop_dict[origins_pop_series[i,1]] = origins_pop_series[i,2]
end

#origins_pop_dict

selected_facilities = PartialSetCoverageProblem("../../../../lima_optimization_output/saved_OD.csv", 900, .8, origins_pop_dict)

#write-out selected_facilities
#writedlm("../../../../lima_optimization_output/selected_facilities_file_from_julia",selected_facilities)
