==> Implementations/FY20/ACC_Peru_Lima_Optimization/JupyterNotebooks/Step6 - Optimization_SetCoverage_Python.ipynb <==
This is a type of location optimization analysis, specifically finding the optimal location of facilites on a network. These are two types of Set-Coverage analysis implemented in **Python**:

### Set-Coverage Problem
#### Objective: Determine the minimum number of facilities and their locations in order to cover all demands within a pre-specified maximum distance (or time) coverage

### Partial Set-Coverage Problem
#### Objective: Determine the minimum number of facilities and their locations in order to cover a given fraction of the population within a pre-specified maximum distance (or time) coverage

more information on GOSTNets Optimization can be found in the wiki: https://github.com/worldbank/GOST_PublicGoods/wiki/GOSTnets-Optimization

import geopandas as gpd
import pandas as pd
import os, sys
# add to your system path the location of the GOSTnet.py library
sys.path.append(r'../../../GOSTNets/GOSTNets')
import GOSTnet_Optimization as gn
import importlib

#### These Set-Coverage Problems require an OD matrix as an input¶
read in OD matrix saved as csv file as dataframe

pth = r'../../../../lima_optimization_output'
OD_df = pd.read_csv(os.path.join(pth, 'saved_OD.csv'),index_col=0)

import pulp

### Run the Set-Coverage Problem
#### Objective: Determine the minimum number of facilities and their locations in order to cover all demands within a pre-specified maximum distance (or time) coverage

In the example below there is a pre-specified maximum distance coverage of 1200 seconds

set_coverage_result = gn.optimize_set_coverage(OD_df, max_coverage = 1200)

set_coverage_result

### Partial Set-Coverage Problem
#### Objective: Determine the minimum number of facilities and their locations in order to cover a given fraction of the population within a pre-specified maximum distance (or time) coverage
This problem factors in population coverage, therefore as an additional input we need to produce a series that has each origin and its respective population

origins_gdf = pd.read_csv(os.path.join(pth, 'origins_snapped.csv'))
#origins_gdf
#origins_gdf.Population.values

origins_w_pop_series = pd.Series(origins_gdf.Population.values, index=origins_gdf.NN)
len(origins_w_pop_series)

#some origins end up snapping to the same nearest node, therefore the code below groups and sums origin populations
origins_w_pop_series_no_dupl = origins_w_pop_series.groupby('NN').sum()
len(origins_w_pop_series_no_dupl)

origins_w_pop_series_no_dupl[:10]

#sum(origins_w_pop_series_no_dupl)
origins_w_pop_series_no_dupl[3]

In the example below, inputs include covering 90 percent of the population, a pre-specified maximum distance coverage of 1200 seconds, and a series of origins with their population

OD_df[:5]

import importlib
importlib.reload(gn)

# neet to double-check this
partial_set_coverage_result = gn.optimize_partial_set_coverage(OD_df, pop_coverage = .8, max_coverage = 900, origins_pop_series = origins_w_pop_series_no_dupl, existing_facilities = None)

partial_set_coverage_result
