==> Implementations/FY20/ACC_Peru_Lima_Optimization/JupyterNotebooks/Step7 - Post optimization - Data prep and mapping.ipynb <==
### Map optimization output back to shapefile
The functions below prepare the data for the process of mapping the results back to their original shapefiles by implementing a "table join" between the outputs .txt and the attribute table of the shapefile. This will allow the visualization of accessibility **after** optimization.
The process for mapping the optimization results is as follows:
    1. we need to filter the snapped destinations file by the IDs from the optimal selected facilities that were generated from the optimization analysis
    2. we need to slice the OD matrix by the new destination file, so that only the columns that represent the optimal selected facilities remain
    3. we map the OD matrix back to the origins shapefile as in Step 4

import pandas as pd
import os, sys
sys.path.append(r'../../../GOSTNets/GOSTNets')
import GOSTnet as gn
import importlib
importlib.reload(gn)
import geopandas as gpd
from shapely.wkt import loads
import numpy as np

# define the paths to the origins, destinations, optimization result
pth = r'../../../../lima_optimization_output'
WGS = {'init':'epsg:4326'}
measure_crs = {'init':'epsg:32718'}

# read in the original destination file (from Step3)
destinations_df = pd.read_csv(os.path.join(pth, 'destinations_snapped.csv'))

destinations_df[:3]

# include the path of the optimization results file *.txt and open the file
Rfile = os.path.join(pth, 'results.txt')
with open (Rfile, 'r') as re:
    results = re.readlines()
results =[(el.strip()) for el in results]
results = [int(i) for i in results]
results[:5]

# filter optimal destinations by the result of the optimization process
destinations = destinations_df[destination_df['NN'].isin(results)]
destinations[:100]

# write optimal destinations file back to .csv for mapping purposes
destinations.to_csv(os.path.join(pth, 'optimal_destinations.csv'))

# read OD matrix
OD = pd.read_csv(os.path.join(pth, r'saved_OD.csv'), sep=',')
OD = OD.rename(columns = {'Unnamed: 0':'NN'})
OD = OD.set_index('NN')

# transform the destination nodes into "string" type to be used as column headers
results_s = [str(x) for x in results]
# filter OD matrix by results
OD = OD[results_s]
OD[:3]

# adding a column that has the minimum time to reach each origin's nearest facility
OD['Minimum'] = OD.loc[:,:].min(axis=1)
OD[:3]

### PREPARE ORIGIN FILE 
# set index to nearest node to match on the travel time from that node to nearest destination (OD['Minimum'])
origins_df = pd.read_csv(os.path.join(pth, 'origins_snapped.csv'))
# orig = orig.rename({'ID':'OBJECTID'}, axis = 1) # rename ID column to 'OBJECTID'
origins_df = origins_df.set_index('NN') # used for matching on the OD matrix
origins_df[:3]

# join works on the index as default 
#join = origins_df.join(OD)

# make sure the index of both dataframes are the same before you try to copy the Minimum column to the origins_df
origins_df['Minimum'] = OD['Minimum']
#origins_df

# add on walktime (kph) for time computation
walkspeed = 4
origins_df['walktime'] = origins_df['NN_dist'] / 1000 / walkspeed * 3600
origins_df['total_time_min'] = (origins_df['walktime'] + origins_df['Minimum']) / 60
###   OR   ###
#compute total distance: NN_dist + dist
#origins_df['total_distance'] = (origins_df['NN_dist'] + origins_df['Minimum'])

# read in the origins in shapefile format
fp = r'../../Peru_Lima_Optimization/test_data'
shpfil = gpd.read_file(os.path.join(fp, 'VillaElSalvador_urban.shp'))

# make sure the index of both dataframes are the same before you try to copy a column over
shpfil = shpfil.set_index('OBJECTID')
origins_df = origins_df.set_index('O_ID')

shpfil['total_time_min'] = origins_df['total_time_min']

# Prep output shapefile, match on visualization column, save
#shpfil['total_time'] = join['total_time']
shpfil = shpfil.to_crs({'init':'epsg:4326'})
shpfil.to_file(os.path.join(pth, 'VES_origins_w_time.shp'), driver = 'ESRI Shapefile')
